remove env.json
