--- packages/web-app/public/config/env.json
@@ -1,11 +1,0 @@
-{
-    "ENV": "dev",
-    "AWS_REGION": "ap-southeast-1",
-    "AWS_COGNITO_IDENDITY_POOL_ID": "ap-southeast-1:6dce3e15-6817-4377-bd82-bb969b254a8f",
-    "AWS_USER_POOLS_ID": "ap-southeast-1_RNMlC2yGY",
-    "AWS_USER_POOLS_WEB_CLIENT_ID": "1om2i227s52mkpl4t256637ba5",
-    "AWS_GRAPHQLENDPOINT_AUTHUSER": "http://localhost:3100/lambda-gq-resolver/auth-user/graphql",
-    "AWS_GRAPHQLENDPOINT_AUTHROLE": "http://localhost:3100/lambda-gq-resolver/auth-role/graphql",
-    "AWS_GRAPHQLENDPOINT_AUTHNONE": "http://localhost:3100/lambda-gq-resolver/auth-none/graphql",
-    "AWS_OATH_DOMAIN": "alpaca-backend-staging.auth.ap-southeast-1.amazoncognito.com"
-}
